+++ src/main/java/spike/commands/ByeCommand.java
@@ -1,3 +1,9 @@
+package spike.commands;
+
+import spike.storage.TaskList;
+import spike.storage.Storage;
+import spike.ui.Ui;
+
 public class ByeCommand extends Command {
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
