+++ src/main/java/spike/commands/ErrorCommand.java
@@ -1,3 +1,9 @@
+package spike.commands;
+
+import spike.storage.TaskList;
+import spike.storage.Storage;
+import spike.ui.Ui;
+
 public class ErrorCommand extends Command {
     private final String message;
