+++ src/main/java/spike/exceptions/SpikeException.java
@@ -1,3 +1,5 @@
+package spike.exceptions;
+
 public class SpikeException extends Exception {
     public SpikeException(String message) {
         super(message);
