+++ src/main/java/spike/parser/Parser.java
@@ -1,3 +1,20 @@
+package spike.parser;
+
+import spike.commands.InputType;
+import spike.commands.AddTaskCommand;
+import spike.commands.ByeCommand;
+import spike.commands.Command;
+import spike.commands.DeleteTaskCommand;
+import spike.commands.ErrorCommand;
+import spike.commands.ListByDateCommand;
+import spike.commands.ListCommand;
+import spike.commands.MarkCommand;
+import spike.commands.UnmarkCommand;
+import spike.tasks.Deadline;
+import spike.tasks.Event;
+import spike.tasks.ToDo;
+import spike.exceptions.SpikeException;
+
 import java.time.DateTimeException;
 import java.time.LocalDateTime;
