+++ src/main/java/spike/tasks/Deadline.java
@@ -1,3 +1,5 @@
+package spike.tasks;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
