+++ src/main/java/spike/tasks/Event.java
@@ -1,3 +1,5 @@
+package spike.tasks;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
