+++ src/main/java/spike/tasks/Task.java
@@ -1,3 +1,5 @@
+package spike.tasks;
+
 public class Task {
     protected String description;
     protected boolean isDone;
