+++ src/main/java/spike/tasks/ToDo.java
@@ -1,3 +1,5 @@
+package spike.tasks;
+
 public class ToDo extends Task {
 
     public ToDo(String description) {
